#0: Reduce the size of tensor_impl.hpp header
