+++ tt_eager/tt_dnn/op_library/groupnorm/groupnorm_op.cpp
@@ -6,8 +6,6 @@
 
 #include <optional>
 
-#include "tt_dnn/op_library/run_operation.hpp"
-#include "tt_eager/tt_dnn/op_library/work_split.hpp"
 #include "tt_eager/tt_dnn/op_library/reshape/reshape_op.hpp"
 #include "tt_eager/tt_dnn/op_library/composite/composite_ops.hpp"
 #include "tt_dnn/op_library/math.hpp"
@@ -15,8 +13,6 @@
 #include "tt_metal/host_api.hpp"
 #include "tt_metal/common/constants.hpp"
 #include "tt_metal/detail/util.hpp"
-
-#include "third_party/magic_enum/magic_enum.hpp"
 
 #include "ttnn/operations/eltwise/binary/binary.hpp"
