+++ tt_eager/tt_dnn/op_library/indexed_fill/indexed_fill_op.cpp
@@ -4,10 +4,6 @@
 
 #include "tt_dnn/op_library/indexed_fill/indexed_fill_op.hpp"
 
-#include "tt_metal/host_api.hpp"
-
-#include "third_party/magic_enum/magic_enum.hpp"
-
 
 namespace tt {
