+++ tt_eager/tt_dnn/op_library/layernorm/layernorm_op.cpp
@@ -3,15 +3,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "tt_eager/tt_dnn/op_library/layernorm/layernorm_op.hpp"
-#include "tt_eager/tt_dnn/op_library/work_split.hpp"
 #include "tt_dnn/op_library/run_operation.hpp"
 #include "tt_dnn/op_library/math.hpp"
 
-#include "tt_metal/host_api.hpp"
 #include "tt_metal/common/constants.hpp"
-#include "tt_metal/detail/util.hpp"
-
-#include "third_party/magic_enum/magic_enum.hpp"
 
 #include <optional>
