+++ tt_eager/tt_dnn/op_library/move/move_op.cpp
@@ -4,10 +4,6 @@
 
 #include "tt_dnn/op_library/move/move_op.hpp"
 
-#include "tt_metal/host_api.hpp"
-
-#include "third_party/magic_enum/magic_enum.hpp"
-
 namespace move_op_utils {
 using namespace tt::tt_metal;
