+++ tt_eager/tt_dnn/op_library/nlp_tms/nlp_tms.cpp
@@ -7,8 +7,6 @@
 
 #include "tt_metal/host_api.hpp"
 
-#include "third_party/magic_enum/magic_enum.hpp"
-
 namespace tt {
 
 namespace tt_metal {
