+++ tt_eager/tt_dnn/op_library/reduce/reduce_op.cpp
@@ -3,17 +3,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "tt_dnn/op_library/reduce/reduce_op.hpp"
+#include "tensor/tensor_utils.hpp"
+#include "tt_dnn/op_library/eltwise_unary/eltwise_unary_op.hpp"
 #include "tt_dnn/op_library/transpose/transpose_op.hpp"
 #include "tt_dnn/op_library/auto_format.hpp"
 #include "tt_dnn/op_library/reshape/reshape_op.hpp"
 #include "tt_dnn/op_library/run_operation.hpp"
-#include "tt_metal/tools/profiler/op_profiler.hpp"
-#include "tt_dnn/op_library/bcast/bcast_op.hpp"
-#include "tt_dnn/op_library/composite/composite_ops.hpp"
 #include "tt_metal/host_api.hpp"
 #include "tt_metal/common/constants.hpp"
 
-#include "third_party/magic_enum/magic_enum.hpp"
 
 #include <limits>
