+++ tt_eager/tt_dnn/op_library/rotary_embedding/rotary_embedding_op.hpp
@@ -7,7 +7,6 @@
 #include <functional>
 
 #include "tensor/tensor.hpp"
-#include "third_party/magic_enum/magic_enum.hpp"
 #include "tt_dnn/op_library/run_operation.hpp"
 #include "tt_metal/host_api.hpp"
 #include "tt_dnn/op_library/compute_kernel_config.hpp"
