+++ tt_eager/tt_dnn/op_library/rotate_half/rotate_half_op.cpp
@@ -4,11 +4,7 @@
 
 #include "tt_dnn/op_library/rotate_half/rotate_half_op.hpp"
 
-#include "tt_metal/host_api.hpp"
 #include "tt_metal/common/constants.hpp"
-#include "tt_metal/host_api.hpp"
-
-#include "third_party/magic_enum/magic_enum.hpp"
 
 using namespace tt::constants;
