+++ tt_eager/tt_dnn/op_library/sliding_window_op_infra/sliding_window.hpp
@@ -7,7 +7,6 @@
 #include <string>
 #include <tuple>
 #include <fmt/core.h>
-#include <fmt/format.h>
 
 #include "tt_eager/tensor/host_buffer/functions.hpp"
 #include "utils.hpp"
