+++ tt_eager/tt_dnn/op_library/transformer_tms/transformer_tms.cpp
@@ -4,10 +4,8 @@
 
 #include "tt_dnn/op_library/transformer_tms/transformer_tms.hpp"
 
-#include "third_party/magic_enum/magic_enum.hpp"
 #include "tt_dnn/op_library/work_split.hpp"
 #include "tt_metal/host_api.hpp"
-#include "tt_metal/tools/profiler/op_profiler.hpp"
 
 namespace tt {
 namespace operations {
