+++ tt_eager/tt_dnn/op_library/update_cache/update_cache_op.hpp
@@ -4,9 +4,6 @@
 
 #pragma once
 
-#include <functional>
-#include "third_party/magic_enum/magic_enum.hpp"
-
 #include "tensor/tensor.hpp"
 #include "tt_dnn/op_library/run_operation.hpp"
 #include "tt_metal/host_api.hpp"
