+++ tt_metal/CMakeLists.txt
@@ -20,6 +20,19 @@ add_library(tt_metal ${TT_METAL_OBJECTS})
 
 target_link_libraries(tt_metal PUBLIC umd_device metal_common_libs)
 
+target_precompile_headers(tt_metal PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/third_party/magic_enum/magic_enum.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/third_party/tracy/public/tracy/Tracy.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/third_party/fmt/fmt/core.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/third_party/fmt/fmt/format.h
+    <functional>
+    <map>
+    <memory>
+    <unordered_map>
+    <variant>
+    <vector>
+)
+
 target_link_libraries(tt_metal PUBLIC compiler_flags linker_flags metal_header_directories $<$<BOOL:$ENV{ENABLE_TRACY}>:TracyClient>)  # linker_flags = -rdynamic if tracy enabled
 target_link_directories(tt_metal PUBLIC ${PROJECT_BINARY_DIR}/lib)    # required so tt_metal can find device library
 target_include_directories(tt_metal PUBLIC
