+++ ttnn/cpp/ttnn/op_library/to_memory_config/to_memory_config_op.hpp
@@ -4,21 +4,12 @@
 
 #pragma once
 
-#include <functional>
 #include <optional>
 
 #include "tensor/tensor.hpp"
-// #include "third_party/magic_enum/magic_enum.hpp"
-// #include "tt_eager/tensor/host_buffer/functions.hpp"
-// #include "tt_eager/tensor/tensor_utils.hpp"
-#include "tt_eager/tt_dnn/op_library/compute_kernel_config.hpp"
 #include "tt_eager/tt_dnn/op_library/copy/copy_op.hpp"
 #include "tt_eager/tt_dnn/op_library/run_operation.hpp"
 #include "tt_eager/tt_dnn/op_library/sharded/sharded_op.hpp"
-// #include "tt_metal/host_api.hpp"
-// #include "tt_metal/impl/dispatch/command_queue.hpp"
-// #include "ttnn/core.hpp"
-#include "ttnn/operations/core.hpp"
 #include "ttnn/types.hpp"
 
 namespace ttnn {
